labs/lab03-foundry-agent/src/requirements.txt
azure-ai-projects>=2.1.0
azure-identity>=1.15.0
openai>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
